requirements.txt
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
cachetools==4.2.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
decorator==5.1.1
frozenlist==1.3.0
future==0.18.2
h11==0.13.0
idna==3.3
multidict==6.0.2
outcome==1.2.0
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
python-whois==0.7.3
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tornado==6.1
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validators==0.20.0
wsproto==1.1.0
yarl==1.7.2
